clean up and move cert-related functions to new cert.{c,h}
Also generalize UTIL_warn_cert{,s} to CERT_check{_all}
