+++ src/util/util.c
@@ -236,151 +236,6 @@ err:
 }
 
 
-/*
- * dn is expected to be in the format /type0=value0/type1=value1/type2=...
- * where characters may be escaped by \
- */
-/* adapted from OpenSSL:apps/lib/apps.c */
-X509_NAME* UTIL_parse_name(const char* dn, long chtype, bool multirdn)
-{
-    size_t buflen = strlen(dn) + 1; /* to copy the types and values.
-                                     * Due to escaping, the copy can only become shorter */
-    char* buf = OPENSSL_malloc(buflen);
-    size_t max_ne = buflen / (1 + 1) + 1; /* maximum number of name elements */
-    const char** ne_types = OPENSSL_malloc(max_ne * sizeof(char*));
-    char** ne_values = OPENSSL_malloc(max_ne * sizeof(char*));
-    int* mval = OPENSSL_malloc(max_ne * sizeof(int));
-
-    const char* sp = dn;
-    char* bp = buf;
-    int i, ne_num = 0;
-
-    X509_NAME* n = 0;
-    int nid;
-
-    if(0 is_eq buf or 0 is_eq ne_types or 0 is_eq ne_values or 0 is_eq mval)
-    {
-        LOG_err("Malloc error");
-        goto error;
-    }
-
-    /* no multivalued RDN by default */
-    mval[ne_num] = 0;
-
-    if(*sp++ not_eq '/')
-    { /* skip leading '/' */
-        LOG(FL_ERR, "DN '%s' does not start with '/'.", dn);
-        goto error;
-    }
-
-    while(*sp not_eq '\0')
-    {
-        /* collect type */
-        ne_types[ne_num] = bp;
-        /* parse element name */
-        while(*sp not_eq '=')
-        {
-            if(*sp is_eq '\\')
-            { /* is there anything to escape in the * type...? */
-                if(*++sp not_eq '\0')
-                {
-                    *bp++ = *sp++;
-                }
-                else
-                {
-                    LOG(FL_ERR, "Escape character at end of DN '%s'", dn);
-                    goto error;
-                }
-            }
-            else if(*sp is_eq '\0')
-            {
-                LOG(FL_ERR, "End of string encountered while processing type of DN '%s' element #%d", dn, ne_num);
-                goto error;
-            }
-            else
-            {
-                *bp++ = *sp++;
-            }
-        }
-        sp++;
-        *bp++ = '\0';
-        /* parse element value */
-        ne_values[ne_num] = bp;
-        while(*sp not_eq '\0')
-        {
-            if(*sp is_eq '\\')
-            {
-                if(*++sp not_eq '\0')
-                {
-                    *bp++ = *sp++;
-                }
-                else
-                {
-                    LOG(FL_ERR, "Escape character at end of DN '%s'", dn);
-                    goto error;
-                }
-            }
-            else if(*sp is_eq '/')
-            { /* start of next element */
-                sp++;
-                /* no multivalued RDN by default */
-                mval[ne_num + 1] = 0;
-                break;
-            }
-            else if(*sp is_eq '+' and multirdn)
-            {
-                /* a not escaped + signals a multi-valued RDN */
-                sp++;
-                mval[ne_num + 1] = -1;
-                break;
-            }
-            else
-            {
-                *bp++ = *sp++;
-            }
-        }
-        *bp++ = '\0';
-        ne_num++;
-    }
-
-    if(0 is_eq(n = X509_NAME_new()))
-    {
-        goto error;
-    }
-
-    for(i = 0; i < ne_num; i++)
-    {
-        if((nid = OBJ_txt2nid(ne_types[i])) is_eq NID_undef)
-        {
-            LOG(FL_WARN, "DN '%s' attribute %s has no known NID, skipped", dn, ne_types[i]);
-            continue;
-        }
-
-        if(0 is_eq * ne_values[i])
-        {
-            LOG(FL_WARN, "No value provided for DN '%s' attribute %s, skipped", dn, ne_types[i]);
-            continue;
-        }
-
-        if(0 is_eq X509_NAME_add_entry_by_NID(n, nid, chtype, (unsigned char*)ne_values[i], -1, -1, mval[i]))
-        {
-            ERR_print_errors(bio_err);
-            LOG(FL_ERR, "Error adding name attribute '/%s=%s'", ne_types[i], ne_values[i]);
-            X509_NAME_free(n);
-            n = 0;
-            goto error;
-        }
-    }
-
-error:
-    OPENSSL_free(ne_values);
-    OPENSSL_free(ne_types);
-    OPENSSL_free(mval);
-    OPENSSL_free(buf);
-    return n;
-}
-
-
 void UTIL_erase_mem(void* dst, size_t len)
 {
     if(dst not_eq 0)
@@ -465,334 +320,6 @@ int ASN1_TIME_compare(const ASN1_TIME *a, const ASN1_TIME *b)
     return 0;
 }
 #endif
-
-/* adatped from OpenSSL:crypto/x509/t_crl.c */
-void UTIL_print_crl(OPTIONAL BIO* bio, OPTIONAL const X509_CRL* crl)
-{
-    unsigned long flags =
-        ASN1_STRFLGS_RFC2253 bitor ASN1_STRFLGS_ESC_QUOTE bitor XN_FLAG_SEP_CPLUS_SPC bitor XN_FLAG_FN_SN;
-
-    if(bio is_eq 0)
-    {
-        return;
-    }
-    if(crl not_eq 0)
-    {
-        BIO_printf(bio, "    Certificate Revocation List (CRL):\n");
-#if 0
-        long ver = X509_CRL_get_version(crl);
-        if(ver >= 0 and ver <= 1)
-        {
-            BIO_printf(bio, "%8sVersion %ld (0x%lx)\n", "", ver + 1, (unsigned long)ver);
-        }
-        else
-        {
-            BIO_printf(bio, "%8sVersion unknown (%ld)\n", "", ver);
-        }
-        const X509_ALGOR* sig_alg;
-        const ASN1_BIT_STRING* sig;
-        X509_CRL_get0_signature(crl, &sig, &sig_alg);
-        BIO_puts(bio, "    ");
-        X509_signature_print(bio, sig_alg, NULL);
-#endif
-        BIO_printf(bio, "%8sIssuer: ", "");
-        X509_NAME_print_ex(bio, X509_CRL_get_issuer(crl), 0, flags);
-        BIO_puts(bio, "\n");
-        BIO_printf(bio, "%8sLast Update: ", "");
-        ASN1_TIME_print(bio, X509_CRL_get0_lastUpdate(crl));
-        BIO_printf(bio, "\n%8sNext Update: ", "");
-        if (X509_CRL_get0_nextUpdate(crl))
-        {
-            ASN1_TIME_print(bio, X509_CRL_get0_nextUpdate(crl));
-        }
-        else
-        {
-            BIO_printf(bio, "NONE");
-        }
-        BIO_printf(bio, "\n");
-
-#if 0
-        X509V3_extensions_print(bio, "CRL extensions",
-                                X509_CRL_get0_extensions(crl), 0, 8);
-#endif
-
-        STACK_OF(X509_REVOKED) * rev = X509_CRL_get_REVOKED((X509_CRL*) crl);
-        BIO_printf(bio, "    Number of revoked Certificates: %d\n",
-                   rev is_eq 0 ? 0 : sk_X509_REVOKED_num(rev));
-
-#if 0
-        int i;
-        for(i = 0; i < sk_X509_REVOKED_num(rev); i++)
-        {
-            X509_REVOKED* r = sk_X509_REVOKED_value(rev, i);
-            BIO_printf(bio, "    Serial Number: ");
-            i2a_ASN1_INTEGER(bio, X509_REVOKED_get0_serialNumber(r));
-            BIO_printf(bio, "\n        Revocation Date: ");
-            ASN1_TIME_print(bio, X509_REVOKED_get0_revocationDate(r));
-            BIO_printf(bio, "\n");
-            X509V3_extensions_print(bio, "CRL entry extensions",
-                                    X509_REVOKED_get0_extensions(r), 0, 8);
-        }
-        X509_signature_print(bio, sig_alg, sig);
-#endif
-    }
-    else
-    {
-        BIO_printf(bio, "    (no CRL)\n");
-    }
-    BIO_flush(bio);
-}
-
-void UTIL_print_cert(OPTIONAL BIO* bio, OPTIONAL const X509* cert, unsigned long neg_cflags)
-{
-    if(bio is_eq 0)
-    {
-        return;
-    }
-    if(cert not_eq 0)
-    {
-        unsigned long flags =
-            ASN1_STRFLGS_RFC2253 bitor ASN1_STRFLGS_ESC_QUOTE bitor XN_FLAG_SEP_CPLUS_SPC bitor XN_FLAG_FN_SN;
-        BIO_printf(bio, "    Certificate\n");
-        X509_print_ex(bio, (X509*)cert, flags, compl X509_FLAG_NO_SUBJECT);
-        if(X509_check_issued((X509*)cert, (X509*)cert) is_eq X509_V_OK)
-        {
-            BIO_printf(bio, "        self-signed\n");
-        }
-        else
-        {
-            BIO_printf(bio, " ");
-            X509_print_ex(bio, (X509*)cert, flags, compl X509_FLAG_NO_ISSUER);
-        }
-        X509_print_ex(bio, (X509*)cert, flags, compl(X509_FLAG_NO_SERIAL bitor X509_FLAG_NO_VALIDITY));
-        if(X509_cmp_current_time(X509_get0_notBefore(cert)) > 0)
-        {
-            BIO_printf(bio, "        not yet valid\n");
-        }
-        if(X509_cmp_current_time(X509_get0_notAfter(cert)) < 0)
-        {
-            BIO_printf(bio, "        no more valid\n");
-        }
-        X509_print_ex(bio, (X509*)cert, flags, compl(neg_cflags));
-    }
-    else
-    {
-        BIO_printf(bio, "    (no certificate)\n");
-    }
-    BIO_flush(bio);
-}
-
-
-void UTIL_print_certs(OPTIONAL BIO* bio, OPTIONAL const STACK_OF(X509) * certs)
-{
-    if(bio is_eq 0)
-    {
-        return;
-    }
-    if(certs and sk_X509_num(certs) > 0)
-    {
-        int i = 0;
-        for(i = 0; i < sk_X509_num(certs); i++)
-        {
-            X509* cert = sk_X509_value(certs, i);
-            if(cert not_eq 0)
-            {
-                UTIL_print_cert(bio, cert, 0);
-            }
-        }
-    }
-    else
-    {
-        BIO_printf(bio, "    (no certificates)\n");
-        BIO_flush(bio);
-    }
-}
-
-/* This is similar to LOG_cert() from ../certstatus/certstatus.c */
-static void warn_cert_msg(const char *uri, X509 *cert, const char *msg)
-{
-    char *subj = X509_NAME_oneline(X509_get_subject_name(cert), NULL, 0);
-
-    if (cert == NULL)
-        LOG_err("null pointer uri argument");
-    if (cert == NULL)
-        LOG_err("null pointer msg argument");
-    LOG(FL_WARN, "Certificate from '%s' with subject '%s' %s\n", uri, subj, msg);
-    OPENSSL_free(subj);
-}
-
-bool UTIL_warn_cert(const char *uri, OPTIONAL X509 *cert, bool warn_EE,
-                    OPTIONAL X509_VERIFY_PARAM *vpm)
-{
-    if (cert == NULL)
-        return true;
-    uint32_t ex_flags = X509_get_extension_flags(cert);
-    int res = 0;
-
-#if OPENSSL_VERSION_NUMBER >= 0x30000000L
-    res = X509_cmp_timeframe(vpm, X509_get0_notBefore(cert),
-                             X509_get0_notAfter(cert));
-#else
-    unsigned long flags = vpm == NULL ? 0 : X509_VERIFY_PARAM_get_flags(vpm);
-    if ((flags & X509_V_FLAG_NO_CHECK_TIME) == 0) {
-        time_t ref_time;
-        time_t *time = NULL;
-        if ((flags & X509_V_FLAG_USE_CHECK_TIME) != 0) {
-            ref_time = X509_VERIFY_PARAM_get_time(vpm);
-            time = &ref_time;
-        }
-        if (X509_cmp_time(X509_get0_notAfter(cert), time) < 0)
-            res = 1;
-        else if (X509_cmp_time(X509_get0_notBefore(cert), time) > 0)
-            res = -1;
-    }
-#endif
-    bool ret = res == 0;
-    if (!ret)
-        warn_cert_msg(uri, cert, res > 0 ? "has expired" : "not yet valid");
-    if (warn_EE && (ex_flags & EXFLAG_V1) == 0 && (ex_flags & EXFLAG_CA) == 0) {
-        warn_cert_msg(uri, cert, "is not a CA cert");
-        ret = false;
-    }
-    return ret;
-}
-
-bool UTIL_warn_certs(const char *uri, OPTIONAL STACK_OF(X509) *certs, bool warn_EE,
-                     OPTIONAL X509_VERIFY_PARAM *vpm)
-{
-    int i;
-    bool ret = true;
-
-    for (i = 0; i < sk_X509_num(certs /* may be NULL */); i++)
-        ret = UTIL_warn_cert(uri, sk_X509_value(certs, i), warn_EE, vpm)
-            && ret; /* Having 'ret' after the '&&', all certs are checked. */
-    return ret;
-}
-
-
-/** get a copy of all certificates in a store */
-STACK_OF(X509) * X509_STORE_get_certs(X509_STORE* store)
-{ /* TODO remove when available in library */
-    int i = 0;
-    STACK_OF(X509) * sk;
-    STACK_OF(X509_OBJECT) * objs;
-
-    if(0 is_eq store)
-    {
-        return 0;
-    }
-    if(0 is_eq(sk = sk_X509_new_null()))
-    {
-        return 0;
-    }
-    objs = X509_STORE_get0_objects(store);
-    for(i = 0; i < sk_X509_OBJECT_num(objs); i++)
-    {
-        X509* cert = X509_OBJECT_get0_X509(sk_X509_OBJECT_value(objs, i));
-        if(cert not_eq 0)
-        {
-            X509_up_ref(cert);
-            if(0 is_eq sk_X509_push(sk, cert))
-            {
-                X509_free(cert);
-                sk_X509_pop_free(sk, X509_free);
-                return 0;
-            }
-        }
-    }
-    return sk;
-}
-
-
-void UTIL_print_store_certs(OPTIONAL BIO* bio, OPTIONAL const X509_STORE* store)
-{
-    if(bio is_eq 0)
-    {
-        return;
-    }
-    if(store not_eq 0)
-    {
-        STACK_OF(X509)* certs = X509_STORE_get_certs((X509_STORE*)store);
-        UTIL_print_certs(bio, certs);
-        sk_X509_pop_free(certs, X509_free);
-    }
-    else
-    {
-        BIO_printf(bio, "    (no certificate store)\n");
-        BIO_flush(bio);
-    }
-}
-
-int UTIL_store_certs_num(const X509_STORE* store)
-{
-    if(store is_eq 0)
-    {
-        return 0;
-    }
-
-    int result = 0;
-    STACK_OF(X509)* certs = X509_STORE_get_certs((X509_STORE*)store);
-    if(certs)
-    {
-        result = sk_X509_num(certs);
-        sk_X509_pop_free(certs, X509_free);
-    }
-
-    return result;
-}
-
-
-/* start of definitions borrowed from OpenSSL:crypto/cmp/cmp_lib.c */
-static int X509_cmp_from_ptrs(const struct x509_st* const* a, const struct x509_st* const* b)
-{
-    return X509_cmp(*a, *b);
-}
-
-bool UTIL_sk_X509_add1_cert(STACK_OF(X509) * sk, X509* cert, bool no_duplicate)
-{
-    if(no_duplicate)
-    {
-        sk_X509_set_cmp_func(sk, &X509_cmp_from_ptrs);
-        if(sk_X509_find(sk, cert) >= 0)
-        {
-            return 1;
-        }
-    }
-    if(0 is_eq sk_X509_push(sk, cert))
-    {
-        return 0;
-    }
-    return X509_up_ref(cert);
-}
-
-int UTIL_sk_X509_add1_certs(STACK_OF(X509) * sk, OPTIONAL const STACK_OF(X509) * certs, int no_self_signed,
-                            int no_duplicates)
-{
-    int i = 0;
-
-    if(sk is_eq 0)
-    {
-        return 0;
-    }
-
-    if(certs is_eq 0)
-    {
-        return 1;
-    }
-    const int n = sk_X509_num(certs);
-    for(i = 0; i < n; i++)
-    {
-        X509* cert = sk_X509_value(certs, i);
-        if((not no_self_signed or X509_check_issued(cert, cert) not_eq X509_V_OK)
-           and (not UTIL_sk_X509_add1_cert(sk, cert, no_duplicates)))
-        {
-            return 0;
-        }
-    }
-    return 1;
-}
-/* end of definitions borrowed from OpenSSL:crypto/cmp/cmp_lib.c */
-
 
 /* Copy a NUL-terminated string from the given source
  * into an optional destination buffer, or calculate how large it needs to be.
